moved redux middleware & related tests to their own modules
