+++ app/src/store.js
@@ -1,11 +1,6 @@
-import thunkMiddleware from "redux-thunk";
 import { createStore, applyMiddleware, compose } from "redux";
 import { rootReducer } from "./reducers";
-import logger from "redux-logger";
-import { logException } from "./utils/logging";
-
-const USE_LOGGER = process.env.USE_REDUX_LOGGER;
-const middlewares = [thunkMiddleware];
+import { middlewares } from "./utils/middleware";
 
 const reduxDevToolsPresent =
   typeof window === "object" && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__;
@@ -14,29 +9,6 @@ const composeEnhancers =
   process.env.NODE_ENV !== "production" && reduxDevToolsPresent
     ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
     : compose;
-
-export const crashReporter = (store) => (next) => (action) => {
-  try {
-    return next(action);
-  } catch (err) {
-    logException(
-      `crashReporter: ${
-        typeof err === "object"
-          ? `${err.name}; ${err.message}; ${JSON.stringify(store.getState())}`
-          : `${err}; ${JSON.stringify(store.getState())}`
-      }`
-    );
-    throw err;
-  }
-};
-
-if (process.env.NODE_ENV === "development" && USE_LOGGER) {
-  middlewares.push(logger);
-}
-
-if (process.env.NODE_ENV === "production") {
-  middlewares.push(crashReporter);
-}
 
 export const store = createStore(
   rootReducer,
