+++ app/src/store.spec.js
@@ -1,45 +1,7 @@
-import { crashReporter, store, observeStore } from "./store";
-import { logException } from "./utils/logging";
-
-jest.mock("./utils/logging");
+import { store, observeStore } from "./store";
 
 const spyGetState = jest.spyOn(store, "getState");
 const spySubscribe = jest.spyOn(store, "subscribe");
-
-describe("crashReporter", () => {
-  const create = () => {
-    const store = {
-      getState: jest.fn(() => ({})),
-      dispatch: jest.fn(),
-    };
-    const next = jest.fn();
-
-    const invoke = (action) => crashReporter(store)(next)(action);
-
-    return { store, next, invoke };
-  };
-
-  test("it passes through non-Errors", () => {
-    const { next, invoke } = create();
-    const action = { type: "TEST" };
-    invoke(action);
-    expect(next).toHaveBeenCalledWith(action);
-  });
-
-  test("it calls logException when an error occurs", () => {
-    const { store, next } = create();
-    next.mockImplementation(() => {
-      throw new Error("Something bad happened");
-    });
-    const action = { type: "TEST" };
-    expect(() => {
-      crashReporter(store)(next)(action);
-    }).toThrow("Something bad happened");
-    expect(logException).toHaveBeenCalledWith(
-      "crashReporter: Error; Something bad happened; {}"
-    );
-  });
-});
 
 describe("observeStore", () => {
   let select;
